requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pandas==2.2.3
scikit-learn==1.5.2
requests==2.32.3
sqlalchemy==2.0.36
pydantic==2.10.3
joblib==1.4.2
numpy>=1.24,<2.0
streamlit==1.40.2
plotly==5.24.1
python-multipart==0.0.12
httpx==0.27.2
python-dotenv>=1.0.0
